This is a broken commit. Protos refactored. None of the service handlers have been changed enough build sucessfully.
