+++ cli/orb/src/subcommand/operator/start.rs
@@ -4,8 +4,8 @@ use structopt::StructOpt;
 use crate::{GlobalOption, SubcommandError};
 
 use orbital_headers::{
-    build_metadata::server::BuildServiceServer, credential::server::CredentialServiceServer,
-    integration::server::IntegrationServiceServer, organization::server::OrganizationServiceServer,
+    build_metadata::server::BuildServiceServer, integration::server::IntegrationServiceServer,
+    organization::server::OrganizationServiceServer, secret::server::SecretServiceServer,
 };
 use orbital_services::OrbitalApi;
 
@@ -33,7 +33,7 @@ pub async fn subcommand_handler(
     debug!("Starting single-node server");
     Server::builder()
         .add_service(BuildServiceServer::new(OrbitalApi::default()))
-        .add_service(CredentialServiceServer::new(OrbitalApi::default()))
+        .add_service(SecretServiceServer::new(OrbitalApi::default()))
         .add_service(IntegrationServiceServer::new(OrbitalApi::default()))
         .add_service(OrganizationServiceServer::new(OrbitalApi::default()))
         .serve(addr)
