+++ models/orbital_headers/src/lib.rs
@@ -1,16 +1,21 @@
-/// Generated Rust from protobufs for build metadata and build services
-pub mod build_metadata {
-    tonic::include_proto!("build_metadata");
+/// Generated Rust from protobufs for build meta functions and build services
+pub mod build_meta {
+    tonic::include_proto!("build_meta");
 }
 
-/// Generated Rust from protobufs for credential handling
-pub mod credential {
-    tonic::include_proto!("credential");
+/// Generated Rust from protobufs for secret handling
+pub mod secret {
+    tonic::include_proto!("secret");
 }
 
-/// Generated Rust from protobufs for external service integration
-pub mod integration {
-    tonic::include_proto!("integration");
+/// Generated Rust from protobufs for external notification integration
+pub mod notify {
+    tonic::include_proto!("notify");
+}
+
+/// Generated Rust from protobufs for source code integration
+pub mod code {
+    tonic::include_proto!("code");
 }
 
 /// Generated Rust from protobufs for high-level units, Organizations
@@ -18,7 +23,7 @@ pub mod organization {
     tonic::include_proto!("organization");
 }
 
-/// Generated Rust from protobufs for possible job states
-pub mod state {
-    tonic::include_proto!("state");
+/// Generated Rust from protobufs for enums used over multiple packages
+pub mod orbital_types {
+    tonic::include_proto!("orbital_types");
 }
