--- models/protos/integration.proto
@@ -1,57 +1,0 @@
-syntax = "proto3";
-package integration;
-
-import "state.proto";
-
-enum NotifyType {
-    NOTIFY_TYPE_UNSPECIFIED = 0;
-    SLACK = 1;
-    WEBHOOK = 2;
-}
-
-// Server returns SecretEntry as response
-message NotifyCreateRequest {
-    string org = 1;
-    NotifyType notify_type = 2;
-    string secret = 3;
-    state.ActiveState active_state = 4;
-}
-
-// Server returns SecretEntry as response
-message NotifyEntry {
-    int32 id = 1;
-    int32 org_id = 2;
-    NotifyType notify_type = 3;
-    state.ActiveState active_state = 4;
-}
-
-// Server returns SecretEntry as response
-message NotifyDeleteRequest {
-    string org = 1;
-    NotifyType notify_type = 2;
-}
-
-// Server returns SecretEntry as response
-message NotifyUpdateRequest {
-    string org = 1;
-    NotifyType notify_type = 2;
-    string secret = 3;
-    state.ActiveState active_state = 4;
-}
-
-// Server returns NotifyListResponse
-message NotifyListRequest {
-    string org = 1;
-    NotifyType notify_type = 2;
-}
-
-message NotifyListResponse {
-    repeated NotifyEntry notify_integration_entries = 1;
-}
-
-service IntegrationService {
-    rpc CreateNotify(NotifyCreateRequest) returns (NotifyEntry);
-    rpc DeleteNotify(NotifyDeleteRequest) returns (NotifyEntry);
-    rpc UpdateNotify(NotifyUpdateRequest) returns (NotifyEntry);
-    rpc ListNotify(NotifyListRequest) returns (NotifyListResponse);
-}
