+++ models/protos/secret.proto
@@ -1,0 +1,64 @@
+syntax = "proto3";
+package secret;
+
+import "orbital_types.proto";
+import "google/protobuf/timestamp.proto";
+
+/// Represents a 
+message SecretEntry {
+    int32 org_name = 1;
+    int32 secret_id = 2;
+    string secret_name = 3;
+    orbital_types.SecretType secret_type = 4;
+    orbital_types.ActiveState active_state = 5;
+    bytes secret_data = 6;
+    google.protobuf.Timestamp modified_time = 7;
+}
+
+// Server returns SecretEntry as response
+message SecretCreateRequest {
+    string org_name = 1;
+    string secret_name = 2;
+    orbital_types.SecretType secret_type = 3;
+    bytes secret_data = 4;
+}
+
+// Server returns SecretEntry as response
+message SecretGetRequest {
+    string org_name = 1;
+    string secret_name = 2;
+    orbital_types.SecretType secret_type = 3;
+}
+
+// Server returns SecretEntry as response
+message SecretUpdateRequest {
+    string org_name = 1;
+    string secret_id = 2;
+    orbital_types.SecretType secret_type = 3;
+    orbital_types.ActiveState active_state = 4;
+}
+
+// Server returns SecretEntry as response
+message SecretDeleteRequest {
+    string org_name = 1;
+    string secret_id = 2;
+    orbital_types.SecretType secret_type = 3;
+}
+
+// Server returns SecretListResponse
+message SecretListRequest {
+    orbital_types.SecretType secret_type = 1;
+}
+
+message SecretListResponse {
+    repeated SecretEntry secret_entries = 1;
+}
+
+service SecretService {
+    rpc SecretAdd(SecretCreateRequest) returns (SecretEntry);
+    rpc SecretGet(SecretGetRequest) returns (SecretEntry);
+    rpc SecretDelete(SecretDeleteRequest) returns (SecretEntry);
+    rpc SecretUpdate(SecretUpdateRequest) returns (SecretEntry);
+    rpc SecretList(SecretListRequest) returns (SecretListResponse);
+
+}
