--- models/protos/state.proto
@@ -1,11 +1,0 @@
-syntax = "proto3";
-package state;
-
-// define entity/capability states
-enum ActiveState {
-    ACTIVE_STATE_UNSPECIFIED = 0;
-    ENABLED = 1;
-    DISABLED = 2;
-    UNKNOWN = 3;
-    DELETED = 100;
-}
